- Bugfix: GMs will no longer generate news related to realm rank changes, and their guilds will no longer gain things. - Changed: Sorry Ryan, I've reactivated the compressed log file deletion, I don't like error.log.gz.gz.gz.gz.gz.gz.gz - Removed: A number of player stats debug messages.
git-svn-id: https://svn.code.sf.net/p/dolserver/code/DOLSharp/trunk@1044 c9877f0d-d2ea-46ae-93cb-121acdf3e971

+++ GameServer/serverproperty/ServerProperties.cs
@@ -431,8 +431,8 @@ namespace DOL.GS.ServerProperties
 		/// <summary>
 		/// Crafting speed bonus in capital cities
 		/// </summary>
-		[ServerProperty("capital_city_crafting_speed_bonus", "Crafting speed bonus in capital cities, 0.5 = 50% increase, 1 = standard", 1)]
-		public static readonly int CAPITAL_CITY_CRAFTING_SPEED_BONUS;
+		[ServerProperty("capital_city_crafting_speed_bonus", "Crafting speed bonus in capital cities, 0.5 = 50% increase, 1 = standard", 1.0)]
+		public static readonly double CAPITAL_CITY_CRAFTING_SPEED_BONUS;
 
 		/// <summary>
 		/// Allow Bounty Points to be gained in Battlegrounds
